skip running if no entries in MODIFIED_CONTENT_FILE
- dirty hack for #51 rpmpush.sh
- dirty hack for #52 puppetpush.sh
- dirty hack for #53 publishcv.sh

+++ puppetpush.sh
@@ -14,6 +14,13 @@ then
     exit ${WORKSPACE_ERR}
 fi
 
+# has anything changed? If yes, then MODIFIED_CONTENT_FILE is not 0 bytes 
+if [[ ! -s "${MODIFIED_CONTENT_FILE}" ]]
+then
+    echo "No entries in ${MODIFIED_CONTENT_FILE} no need to continue with $0"
+    exit 0
+fi
+
 # Refresh the PULP_MANIFEST
 cd ${PUPPET_REPO}
 rm PULP_MANIFEST
